requirements.txt
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.0.0
reportlab>=4.0.0
